+++ kifi-backend/modules/common/app/com/keepit/common/service/ServiceClient.scala
@@ -1,5 +1,6 @@
 package com.keepit.common.service
 
+import scala.collection.mutable.{ SynchronizedSet, HashSet, Set }
 import scala.concurrent.Future
 import scala.util.Random
 
@@ -9,7 +10,7 @@ import com.keepit.common.logging.Logging
 import com.keepit.common.net.{ CallTimeouts, ClientResponse, HttpClient, HttpUri }
 import com.keepit.common.routes._
 import com.keepit.common.zookeeper.{ ServiceCluster, ServiceInstance }
-import com.keepit.common._
+import com.keepit.common.core._
 import com.keepit.common.strings._
 
 import java.net.ConnectException
@@ -17,12 +18,15 @@ import java.net.ConnectException
 import play.api.libs.json.{ JsNull, JsValue }
 import play.api.libs.concurrent.Execution.Implicits._
 import com.keepit.common.routes.ServiceRoute
+import scala.collection.mutable.{ SynchronizedSet, HashSet, Set => MutableSet, ListBuffer }
 
 class ServiceNotAvailableException(serviceType: ServiceType)
   extends Exception(s"Service of type ${serviceType.name} is not available")
 
 object ServiceClient {
   val MaxUrlLength = 1000
+  sealed trait Register[V] extends MutableSet[V]
+  final class HashSetRegister[V] extends HashSet[V] with SynchronizedSet[V] with Register[V]
   val register = new HashSetRegister[ServiceClient]
 }
