+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/website/HomeController.scala
@@ -3,7 +3,7 @@ package com.keepit.controllers.website
 import com.google.inject.Inject
 
 import com.keepit.commanders.{ InviteCommander, UserCommander, LocalUserExperimentCommander }
-import com.keepit.common.KestrelCombinator
+import com.keepit.common.core._
 import com.keepit.common.akka.SafeFuture
 import com.keepit.common.controller.{ ShoeboxServiceController, ActionAuthenticator, AuthenticatedRequest, WebsiteController }
 import com.keepit.common.db.ExternalId
